pyproject.toml
[project]
name = "card-fraud-e2e-load-testing"
version = "0.1.0"
description = "E2E and load testing suite for card fraud detection services"
requires-python = ">=3.14"
dependencies = [
    # Core
    "locust==2.29.1",
    "httpx>=0.27",
    "pydantic>=2.10",
    # Data generation
    "faker>=33.0",
    # Testing
    "pytest>=8.3",
    "pytest-html>=4.1",
    "pytest-cov>=5.0",
    # Reporting
    "jinja2>=3.1",
    "pandas>=2.0",
    # Utilities
    "python-dateutil>=2.8",
    "pyyaml>=6.0",
    # Auth
    "python-jose[cryptography]>=3.3.0",
    "boto3>=1.42.39",
]

[project.optional-dependencies]
load-test = [
    "locust==2.29.1",
]
dev = [
    "ruff>=0.8",
    "pre-commit>=3.6",
]

[project.scripts]
# Load testing commands
lt-run = "scripts.run_load_test:main"
lt-web = "scripts.run_load_test:web_ui"
lt-rule-engine = "scripts.run_load_test:cli_rule_engine"
lt-rule-mgmt = "scripts.run_load_test:cli_rule_management"
lt-trans-mgmt = "scripts.run_load_test:cli_transaction_management"

# Data generation
gen-users = "scripts.generate_users:main"
gen-transactions = "scripts.generate_transactions:main"
gen-rules = "scripts.generate_rules:main"
gen-rulesets = "scripts.generate_rulesets:main"

# Reporting
gen-report = "scripts.generate_report:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = [
    "--import-mode=prepend",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html:html-reports/coverage",
    "--html=html-reports/pytest/index.html",
    "--self-contained-html",
]
filterwarnings = ["ignore::DeprecationWarning"]
markers = [
    "unit: Unit tests for generators",
    "integration: Integration tests for task sets",
    "load: Load test scenarios",
]

[tool.ruff]
line-length = 100
[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[dependency-groups]
dev = [
    "ruff>=0.14.14",
]

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "scripts"]
